+++ components/createSection.mjs
@@ -1,0 +1,7 @@
+const createSection = (headingSize, headingText, children) =>
+  `<section>
+    <h${headingSize}>${headingText}</h${headingSize}>
+    ${children}
+  </section>`;
+
+export default createSection;
