+++ components/output/displayFibonacciNums.mjs
@@ -1,0 +1,3 @@
+const displayFibonacciNums = (fibonacciNumbers) => `<p>${fibonacciNumbers}</p>`;
+
+export default displayFibonacciNums;
