+++ events/event-handlers/handleFibonacciForm.mjs
@@ -1,0 +1,3 @@
+const handleFibonacciForm = (event) => event.preventDefault();
+
+export { handleFibonacciForm };
