+++ events/event-listeners/setEventListener.mjs
@@ -1,0 +1,4 @@
+const setEventListener = (targetName, eventName, eventHandler) =>
+  document.querySelector(targetName).addEventListener(eventName, eventHandler);
+  
+export { setEventListener };
