+++ logic/calcFibonacciNums.mjs
