--- modules/events/event-handlers/handleFibonacciForm.mjs
@@ -1,6 +1,0 @@
-const handleFibonacciForm = (event) => {
-  event.preventDefault();
-  console.log("The form is submitted!");
-};
-
-export { handleFibonacciForm };
