--- modules/events/event-listeners/setEventListener.mjs
@@ -1,7 +1,0 @@
-const setEventListener = (targetName, eventName, eventHandler) => {
-  return document
-    .querySelector(targetName)
-    .addEventListener(eventName, eventHandler);
-};
-
-export { setEventListener };
